dissolve core; group iss in new subpackage
