+++ fruits/__init__.py
@@ -1,12 +1,5 @@
-from fruits.core.fruit import Fruit, FruitBranch
-from fruits.core.builder import build
-from fruits.core import callback
-
-from fruits.iss import ISS
-
-from fruits import cache
-from fruits import core
-from fruits import iss
-from fruits import preparation
-from fruits import sieving
-from fruits import words
+from . import cache, callback, iss, preparation, seed, sieving
+from .builder import build
+from .fruit import Fruit, FruitBranch
+from .iss import words
+from .iss.iss import ISS
