+++ fruits/iss/words/__init__.py
@@ -1,0 +1,4 @@
+from . import letters
+from .creation import of_weight, replace_letters
+from .letters import ExtendedLetter, letter
+from .word import SimpleWord, Word
