+++ fruits/preparation/__init__.py
@@ -1,5 +1,4 @@
-from fruits.preparation.abstract import Preparateur
-
-from fruits.preparation.dimension import *
-from fruits.preparation.transform import *
-from fruits.preparation.filter import *
+from .abstract import Preparateur
+from .dimension import *
+from .filter import *
+from .transform import *
