+++ fruits/seed.py
@@ -1,9 +1,7 @@
 from abc import ABC, abstractmethod
 from typing import TypeVar
-import warnings
 
 import numpy as np
-
 
 TCopy = TypeVar("TCopy", bound="Seed")
 
@@ -38,30 +36,3 @@ class Seed(ABC):
 
     def __str__(self) -> str:
         return self.__class__.__name__
-
-
-def force_input_shape(X: np.ndarray) -> np.ndarray:
-    """Makes the attempt to format the input shape of the given
-    multidimensional time series dataset.
-    This leads to an three dimensional array where
-
-    - ``X.shape[0]``: Number of time series
-    - ``X.shape[1]``: Number of dimensions in each time series
-    - ``X.shape[2]``: Length of each time series
-
-    Raises:
-        ValueError: If ``X.ndim > 3``.
-    """
-    if X.ndim < 3:
-        warnings.warn("The input shape of the time series dataset "
-                      "is < 3 and needs to be converted.")
-        if X.ndim == 1:
-            X = np.expand_dims(X, axis=0)
-        if X.ndim == 2:
-            X = np.expand_dims(X, axis=1)
-        return X
-    elif X.ndim == 3:
-        return X.copy()
-    else:
-        raise ValueError("Unsupported input shape of the given time series "
-                         "dataset")
