--- fruits/words/__init__.py
@@ -1,5 +1,0 @@
-from fruits.words.word import Word, SimpleWord
-from fruits.words.letters import ExtendedLetter, letter
-from fruits.words.creation import of_weight, replace_letters
-
-from fruits.words import letters
